lifeaid_backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
python-dotenv==1.0.1
PyMySQL==1.1.1
razorpay==1.4.2
twilio==9.3.2
Pillow==10.4.0
google-genai==1.73.1
